Use maximum weight matching algorithm.
The program first triangulates the domain. However, we want to have quad elements.
In order to do that we should omit some edges between the triangles to form a quad. The problem is that we don't know which edges to omit.
To find the worst edges that could be omit the graph theory is used. Every face element is added as vertex and the edges between them becomes the edge of the graph.
Every edge of the graph has weight, the one that could be omitted has higher edge weight and the ones that can't be omit has lower edge weight. Maximum weight matching algorithm
finds the matchings which has the highest weight and the matched edges are omit. This algorithm doesn't promise to generate %100 quads there could be triangles according to the defects in the domain.